requirements.txt
affine==2.4.0
anyio==3.7.0
asyncpg==0.27.0
attrs==23.1.0
certifi==2023.5.7
click==8.1.3
click-plugins==1.1.1
cligj==0.7.2
databases==0.6.2
environs==9.5.0
exceptiongroup==1.1.1
fastapi==0.96.0
greenlet==2.0.2
h11==0.14.0
idna==3.4
marshmallow==3.19.0
numpy==1.24.3
opencv-python==4.7.0.72
ormar==0.12.1
packaging==23.1
pandas==2.0.2
Pillow==9.5.0
pydantic==1.10.4
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3
rasterio==1.3.7
sa==0.0.4
scipy==1.10.1
sharedmem==0.3.8
six==1.16.0
sniffio==1.3.0
snuggs==1.4.7
SQLAlchemy==1.4.41
starlette==0.27.0
typing_extensions==4.6.3
tzdata==2023.3
uvicorn==0.22.0
